refactor(expo-polyfills): move all storage stuff into `storage.ts`
